ModuleDocstrings

Documentation for ModuleDocstrings.

ModuleDocstrings.ModuleDocstrings

ModuleDocstrings.generate

ModuleDocstrings.write

ModuleDocstrings.write

ModuleDocstrings.ModuleDocstrings — Module

ModuleDocstrings.generate: Create an API summary docstring for a module.

ModuleDocstrings.write: add an API summary docstring to a package.

ModuleDocstrings.generate — Method

ModuleDocstrings.generate(mod::Module)

Return an API summary string for mod.

The summary is assembled from all docstrings in the package, picking the first sentence of each docstring. Expect to need to edit these by hand to produce something truly useful.

ModuleDocstrings.write — Method

ModuleDocstrings.write(mod, str)

Edit the module-defining file to insert str as a docstring for mod.

The package should be checked out in develop mode before calling write.

ModuleDocstrings.write — Method

ModuleDocstrings.write(mod)

Modify the source file for mod to add an API summary docstring.

The docstring is produced by ModuleDocstrings.generate. The package should be checked out in develop mode before calling write.